screenshot image color management

+++ src/Starward/Features/Screenshot/ScreenshotHelper.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
@@ -39,6 +40,10 @@ internal static class ScreenshotHelper
     }
 
 
+    private static string avifdec = Path.Combine(AppContext.BaseDirectory, "avifdec.exe");
+
+    private static string djxl = Path.Combine(AppContext.BaseDirectory, "djxl.exe");
+
 
     public static async Task<string> ConvertToJpgAsync(string filePath)
     {
@@ -49,13 +54,56 @@ internal static class ScreenshotHelper
         string cacheFolder = Path.Combine(AppConfig.CacheFolder, "cache");
         Directory.CreateDirectory(cacheFolder);
         string jpgFilePath = Path.Combine(cacheFolder, Path.GetFileNameWithoutExtension(filePath) + ".jpg");
-        using var fs = File.OpenRead(filePath);
-        var decoder = await BitmapDecoder.CreateAsync(fs.AsRandomAccessStream());
-        var ms = new MemoryStream();
-        var encoder = await BitmapEncoder.CreateAsync(BitmapEncoder.JpegEncoderId, ms.AsRandomAccessStream());
-        encoder.SetSoftwareBitmap(await decoder.GetSoftwareBitmapAsync());
-        await encoder.FlushAsync();
-        await File.WriteAllBytesAsync(jpgFilePath, ms.ToArray());
+        string extension = Path.GetExtension(filePath).ToLowerInvariant();
+
+        if (extension is ".avif" && File.Exists(avifdec))
+        {
+            Process? process = Process.Start(new ProcessStartInfo
+            {
+                FileName = avifdec,
+                Arguments = $"""
+                      "{filePath}" "{jpgFilePath}" -q 90
+                      """,
+                CreateNoWindow = true,
+                RedirectStandardOutput = true,
+                RedirectStandardError = true,
+            }) ?? throw new NullReferenceException("avifdec process is null");
+            await process.WaitForExitAsync();
+            if (process.ExitCode != 0 || !File.Exists(jpgFilePath))
+            {
+                string error = await process.StandardError.ReadToEndAsync();
+                throw new Exception($"avifdec failed: {error}");
+            }
+        }
+        else if (extension is ".jxl" && File.Exists(djxl))
+        {
+            Process? process = Process.Start(new ProcessStartInfo
+            {
+                FileName = djxl,
+                Arguments = $"""
+                      "{filePath}" "{jpgFilePath}" -q 90
+                      """,
+                CreateNoWindow = true,
+                RedirectStandardOutput = true,
+                RedirectStandardError = true,
+            }) ?? throw new NullReferenceException("djxl process is null");
+            await process.WaitForExitAsync();
+            if (process.ExitCode != 0 || !File.Exists(jpgFilePath))
+            {
+                string error = await process.StandardError.ReadToEndAsync();
+                throw new Exception($"djxl failed: {error}");
+            }
+        }
+        else
+        {
+            using var fs = File.OpenRead(filePath);
+            var decoder = await BitmapDecoder.CreateAsync(fs.AsRandomAccessStream());
+            var ms = new MemoryStream();
+            var encoder = await BitmapEncoder.CreateAsync(BitmapEncoder.JpegEncoderId, ms.AsRandomAccessStream());
+            encoder.SetSoftwareBitmap(await decoder.GetSoftwareBitmapAsync());
+            await encoder.FlushAsync();
+            await File.WriteAllBytesAsync(jpgFilePath, ms.ToArray());
+        }
         return jpgFilePath;
     }
